added to jacoco for ignoring coverage, moved a util
