+++ app/src/main/java/com/kickstarter/libs/utils/SafeLet.kt
@@ -1,4 +1,4 @@
-package com.kickstarter.ui.extensions
+package com.kickstarter.libs.utils
 
 inline fun <T1 : Any, T2 : Any, R : Any> safeLet(p1: T1?, p2: T2?, block: (T1, T2) -> R?): R? {
     return if (p1 != null && p2 != null) block(p1, p2) else null
