+++ app/src/main/java/com/kickstarter/ui/compose/designsystem/KSAlerts.kt
@@ -25,9 +25,9 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.Popup
+import com.kickstarter.libs.utils.safeLet
 import com.kickstarter.ui.compose.designsystem.KSTheme.colors
 import com.kickstarter.ui.compose.designsystem.KSTheme.typography
-import com.kickstarter.ui.extensions.safeLet
 
 // Snackbars
 @Composable
